+++ src/MinFraud/Model/Insights.php
@@ -6,13 +6,23 @@ namespace MaxMind\MinFraud\Model;
  * Class Insights
  * @package MaxMind\MinFraud\Model
  *
- * @property \MaxMind\MinFraud\Model\CreditCard $creditCard An object containing
- * minFraud data about the credit card used in the transaction.
- * @property \MaxMind\MinFraud\Model\IpAddress $ipAddress An object containing
- * GeoIP2 and minFraud Insights information about the geolocated IP address.
  * @property \MaxMind\MinFraud\Model\BillingAddress $billingAddress An object
  * containing minFraud data related to the billing address used in the
  * transaction.
+ *
+ * @property \MaxMind\MinFraud\Model\CreditCard $creditCard An object containing
+ * minFraud data about the credit card used in the transaction.
+ *
+ * @property \MaxMind\MinFraud\Model\Device $device This object contains
+ * information about the device that MaxMind believes is associated with the
+ * IP address passed in the request.
+ *
+ * @property \MaxMind\MinFraud\Model\Email $email This object contains
+ * information about the email address passed in the request.
+ *
+ * @property \MaxMind\MinFraud\Model\IpAddress $ipAddress An object containing
+ * GeoIP2 and minFraud Insights information about the geolocated IP address.
+ *
  * @property \MaxMind\MinFraud\Model\ShippingAddress $shippingAddress An object
  * containing minFraud data related to the shipping address used in the
  * transaction.
@@ -22,7 +32,7 @@ class Insights extends Score
     /**
      * @internal
      */
-    protected $ipAddress;
+    protected $billingAddress;
 
     /**
      * @internal
@@ -32,23 +42,38 @@ class Insights extends Score
     /**
      * @internal
      */
-    protected $shippingAddress;
+    protected $device;
 
     /**
      * @internal
      */
-    protected $billingAddress;
+    protected $email;
+
+    /**
+     * @internal
+     */
+    protected $ipAddress;
+
+    /**
+     * @internal
+     */
+    protected $shippingAddress;
 
     public function __construct($response, $locales = ['en'])
     {
         parent::__construct($response, $locales);
-        $this->ipAddress
-            = new IpAddress($this->safeArrayLookup($response['ip_address']), $locales);
-        $this->creditCard
-            = new CreditCard($this->safeArrayLookup($response['credit_card']));
-        $this->shippingAddress
-            = new ShippingAddress($this->safeArrayLookup($response['shipping_address']));
+
         $this->billingAddress
             = new BillingAddress($this->safeArrayLookup($response['billing_address']));
+        $this->creditCard
+            = new CreditCard($this->safeArrayLookup($response['credit_card']));
+        $this->device
+            = new Device($this->safeArrayLookup($response['device']));
+        $this->email
+            = new Email($this->safeArrayLookup($response['email']));
+        $this->ipAddress
+            = new IpAddress($this->safeArrayLookup($response['ip_address']), $locales);
+        $this->shippingAddress
+            = new ShippingAddress($this->safeArrayLookup($response['shipping_address']));
     }
 }